+++ KitCli.Playground.App.Terminal/TestCliCommand.cs
@@ -1,0 +1,5 @@
+using KitCli.Commands.Abstractions;
+
+namespace KitCli.Playground.App.Terminal;
+
+public record TestCliCommand : CliCommand;